refactor: 이슈 관련 페이지 폴더 구조 변경 및 이슈 상세 컴포넌트 분리
