+++ frontend/src/components/Main.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import Header from "./header/Header";
 import Filter from "./filter/Filter";
-import IssueList from "./issue/IssueList";
+import IssueList from "./issue/list/IssueList";
 
 function Main() {
   return (
