+++ frontend/src/router/router.tsx
@@ -3,8 +3,8 @@ import useUserStore from "../hooks/stores/useUserStore";
 import Login from "../components/login/Login";
 import Registration from "../components/registration/Registration";
 import Main from "../components/Main";
-import IssueCreator from "../components/creator/IssueCreator";
-import IssueDetail from "../components/issue/IssueDetail";
+import IssueCreator from "../components/issue/new/IssueCreator";
+import IssueDetail from "../components/issue/detail/IssueDetail";
 import LabelList from "../components/label/LabelList";
 import MilestoneList from "../components/milestone/MilestoneList";
